install: add --ignition-url
This matches our support for it for the disk image.

A big difference though is that we don't usually have GPG signatures for
integrity checking and authenticity. At the very least, we refuse to
fetch over pure HTTP without passing `--ignition-hash` or
`--insecure-ignition`. (I didn't want to wrap this under the same
`--insecure` switch since one may still want to verify signatures on the
disk image itself).

Closes: #118