WIP/RFC: Add `coreos-installer iso uproot`
This command takes a live ISO and generates from it a minimal version
which does not contain the root squashfs. The use case for minimal ISOs
has come up multiple times, notably in:

https://github.com/coreos/fedora-coreos-tracker/issues/661

This is a sort of intermediate approach where we don't officially ship a
new artifact, but allow users to derive the minimal ISO from the
official live ISO.

The way this works is similar to osmet. It requires a "packing" step at
build time which adds an osmet-like file (the uproot file) on the full
ISO itself. The uproot file describes how to construct a minimal ISO
from the full ISO. It weighs in at a few kilobytes, so the impact of the
size of the live ISO is negligible. If we ever officially do ship a
minimal ISO, it also allows for the possibility of matching its checksum
exactly.

The method should be arch-independent and does not rely on e.g.
modifying ISO9660 structures or moving the GPT backup header, etc...

Example on the compose side (i.e. in `cosa buildextend-live`):

```
$ ls -lh *.iso*
-rw-r--r--. 1 jlebon jlebon 940M Jun  5 11:33 fedora-coreos-34.20210603.dev.0-live.x86_64.iso
-rw-r--r--. 1 jlebon jlebon  89M Jun  4 17:45 fedora-coreos-34.20210603.dev.0-live.x86_64.iso.minimal

$ # pack the minimal ISO into the live ISO
$ coreos-installer iso uproot-pack *.iso*
```

On the client side:

```
$ ls -lh *.iso
-rw-r--r--. 1 jlebon jlebon 940M Jun  5 11:33 fedora-coreos-34.20210603.dev.0-live.x86_64.iso

$ coreos-installer iso uproot --output minimal.iso --save-rootfs rootfs.img \
        fedora-coreos-34.20210603.dev.0-live.x86_64.iso

$ ls -lh *.iso rootfs.img
-rw-r--r--. 1 jlebon jlebon 940M Jun  5 11:33 fedora-coreos-34.20210603.dev.0-live.x86_64.iso
-rw-------. 1 jlebon jlebon  89M Jun 11 12:42 minimal.iso
-rw-r--r--. 1 jlebon jlebon 849M Jun 11 12:42 rootfs.img
```

Verifying the hashes match:

```
$ sha256sum fedora-coreos-34.20210603.dev.0-live.x86_64.iso.minimal minimal.iso \
        fedora-coreos-34.20210603.dev.0-live-rootfs.x86_64.img rootfs.img
0649c6600c9b1730994e2fad08ad3b57d08c699d04806ffb600acb32921712bf fedora-coreos-34.20210603.dev.0-live.x86_64.iso.minimal
0649c6600c9b1730994e2fad08ad3b57d08c699d04806ffb600acb32921712bf minimal.iso
733b938d6e7fbf0efae07b61cb405d09de3f757eef21c865d3c36e99adb6bae4 fedora-coreos-34.20210603.dev.0-live-rootfs.x86_64.img
733b938d6e7fbf0efae07b61cb405d09de3f757eef21c865d3c36e99adb6bae4 rootfs.img
```

We'll want to also add a `--rootfs-url` which uses the `iso kargs` code
to inject `coreos.live.rootfs_url` since that's the primary use case so
it can all be done in a single step.

One of the primary goals is to ideally satisfy the needs of the Assisted
Installer (https://github.com/openshift/assisted-installer), though
there's more things needed before we get there (notably #545).

WARNING: The code here is still really rough, and does some hacky
things. If the general idea is appealing, I'll clean things up. Might
also use https://github.com/ids1024/iso9660-rs instead of doing the
ISO9660 parsing ourselves.